TWA-4 updated folder structure, client integration complete
