+++ client/src/components/user/GoogleSignIn.jsx
@@ -1,53 +1,44 @@
 import React from 'react'
-import { initializeApp } from "firebase/app";
-import { GoogleAuthProvider,getAuth,signInWithPopup } from "firebase/auth";
-
-
+import { GoogleAuthProvider, getAuth, signInWithPopup } from "firebase/auth";
+import firebase from '../../config/firebase'
 
 function GoogleSignIn() {
+  const initializeGoogleSignIn = () => {
+    const provider = new GoogleAuthProvider();
+    provider.addScope('profile');
+    provider.addScope('email');
+    const auth = getAuth(firebase);
 
-  const provider = new GoogleAuthProvider();
-  provider.setCustomParameters({
-    'login_hint': 'user@example.com'
-  });
-
-  const firebaseConfig = {
-    apiKey: "AIzaSyB8FzeFOgkUgc9KfhaTEYOtBp7jA1xdZFY",
-    authDomain: "theora-web-app.firebaseapp.com",
-    projectId: "theora-web-app",
-    storageBucket: "theora-web-app.appspot.com",
-    messagingSenderId: "82276966171",
-    appId: "1:82276966171:web:3c5e109df0f4aa164b7dbe",
-    measurementId: "G-2TLCCFXMC2"
-  };
-  
-  // Initialize Firebase
-  const app = initializeApp(firebaseConfig);
-
-  const auth = getAuth(app);
-signInWithPopup(auth, provider)
-  .then((result) => {
-    // This gives you a Google Access Token. You can use it to access the Google API.
-    const credential = GoogleAuthProvider.credentialFromResult(result);
-    console.log(credential)
-    const token = credential.accessToken;
-    // The signed-in user info.
-    const user = result.user;
-    // IdP data available using getAdditionalUserInfo(result)
-    // ...
-  }).catch((error) => {
-    // Handle Errors here.
-    const errorCode = error.code;
-    const errorMessage = error.message;
-    // The email of the user's account used.
-    const email = error.customData.email;
-    // The AuthCredential type that was used.
-    const credential = GoogleAuthProvider.credentialFromError(error);
-    // ...
-  }); 
+    signInWithPopup(auth, provider)
+      .then((result) => {
+        // This gives you a Google Access Token. You can use it to access the Google API.
+        const credential = GoogleAuthProvider.credentialFromResult(result);
+        console.log(result)
+        const token = credential.accessToken;
+        // The signed-in user info.
+        const user = result.user;
+        console.log(user)
+        // IdP data available using getAdditionalUserInfo(result)
+        // ...
+      }).catch((error) => {
+        // Handle Errors here.
+        const errorCode = error.code;
+        const errorMessage = error.message;
+        // The email of the user's account used.
+        const email = error.customData.email;
+        // The AuthCredential type that was used.
+        const credential = GoogleAuthProvider.credentialFromError(error);
+        console.log(error)
+        // ...
+      });
+  }
 
   return (
-    <button type="button" className="text-white bg-[#4285F4] hover:bg-[#4285F4]/90 focus:ring-4 focus:outline-none focus:ring-[#4285F4]/50 font-medium rounded-lg text-sm px-5 py-2.5 text-center inline-flex items-center dark:focus:ring-[#4285F4]/55 mr-2 mb-2">
+    <button
+      type="button"
+      className="text-white bg-[#4285F4] hover:bg-[#4285F4]/90 focus:ring-4 focus:outline-none focus:ring-[#4285F4]/50 font-medium rounded-lg text-sm px-5 py-2.5 text-center inline-flex items-center dark:focus:ring-[#4285F4]/55 mr-2 mb-2"
+      onClick={initializeGoogleSignIn}
+    >
       <svg className="w-4 h-4 mr-2" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 18 19">
         <path fillRule="evenodd" d="M8.842 18.083a8.8 8.8 0 0 1-8.65-8.948 8.841 8.841 0 0 1 8.8-8.652h.153a8.464 8.464 0 0 1 5.7 2.257l-2.193 2.038A5.27 5.27 0 0 0 9.09 3.4a5.882 5.882 0 0 0-.2 11.76h.124a5.091 5.091 0 0 0 5.248-4.057L14.3 11H9V8h8.34c.066.543.095 1.09.088 1.636-.086 5.053-3.463 8.449-8.4 8.449l-.186-.002Z" clipRule="evenodd" />
       </svg>
